+++ python/GafferSceneUI/__init__.py
@@ -158,6 +158,7 @@ import CopyPrimitiveVariablesUI
 import MergeScenesUI
 import ShuffleAttributesUI
 import ShufflePrimitiveVariablesUI
+import LocaliseAttributesUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.