+++ Changes.md
@@ -8,6 +8,7 @@ Features
   - The current EditScope may be chosen from a new menu in the Viewer toolbar.
   - The TransformTools and CameraTool will now automatically create edits in the chosen EditScope, enabling edits without the need to manually create a Transform node.
   - Added a `Delete` hotkey which prunes the currently selected objects using the current EditScope.
+- LocaliseAttributes : Added new node which makes local copies of inherited attributes.
 
 Improvements
 ------------